Revert "Mixed bag: Update compatability + depreceations + bugfixes + OIDC enhancements"
