+++ Docker/.env.example
@@ -92,9 +92,6 @@ OIDC_ISSUER_URL=
 # OIDC scope (space-separated list of scopes)
 OIDC_SCOPE=openid email profile
 
-# OIDC admin group (optional, requires group scope)
-OIDC_ADMIN_GROUP=
-
 ### **Development/Debugging Configuration (Optional)**
 
 # Flask environment (development/production)
