+++ backend/auth_utils.py
@@ -1,6 +1,6 @@
 # backend/auth_utils.py
 import jwt
-from datetime import datetime, UTC, timedelta
+from datetime import datetime, timedelta
 from flask import current_app, request, jsonify
 from functools import wraps
 import re
@@ -14,9 +14,9 @@ except ImportError:
 def generate_token(user_id):
     """Generate a JWT token for the user"""
     payload = {
-        'exp': datetime.now(UTC) + current_app.config['JWT_EXPIRATION_DELTA'],
-        'iat': datetime.now(UTC),
-        'sub': str(user_id)
+        'exp': datetime.utcnow() + current_app.config['JWT_EXPIRATION_DELTA'],
+        'iat': datetime.utcnow(),
+        'sub': user_id
     }
     return jwt.encode(payload, current_app.config['SECRET_KEY'], algorithm='HS256')
 
@@ -67,7 +67,7 @@ def token_required(f):
             with conn.cursor() as cur:
                 # Try to get user with is_owner column, fall back to without it if column doesn't exist
                 try:
-                    cur.execute('SELECT id, username, email, is_admin, is_owner, oidc_sub FROM users WHERE id = %s AND is_active = TRUE', (user_id,))
+                    cur.execute('SELECT id, username, email, is_admin, is_owner FROM users WHERE id = %s AND is_active = TRUE', (user_id,))
                     user = cur.fetchone()
                     has_owner_column = True
                 except Exception as e:
@@ -87,13 +87,8 @@ def token_required(f):
                     'username': user[1],
                     'email': user[2],
                     'is_admin': user[3],
+                    'is_owner': user[4] if has_owner_column and len(user) > 4 else False
                 }
-
-                if has_owner_column:
-                    request.user.update({
-                        'is_owner': user[4],
-                        'oidc_managed': user[5] is not None
-                    })
                 
                 return f(*args, **kwargs)
         except Exception as e:
