+++ backend/notifications.py
@@ -12,7 +12,7 @@ import time
 import atexit
 import smtplib
 import logging
-from datetime import datetime, date, UTC, timedelta
+from datetime import datetime, date, timedelta
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 from threading import Lock
@@ -299,8 +299,6 @@ def process_email_notifications(all_warranties, eligible_user_ids, is_manual, ge
     smtp_port = int(os.environ.get('SMTP_PORT', '1025'))
     smtp_username = os.environ.get('SMTP_USERNAME', 'notifications@warracker.com')
     smtp_password = os.environ.get('SMTP_PASSWORD', '')
-    if os.environ.get('SMTP_PASSWORD_FILE'):
-        smtp_password = open(os.environ.get('SMTP_PASSWORD_FILE'), 'r').read().strip()
     smtp_use_tls_env = os.environ.get('SMTP_USE_TLS', 'not_set').lower()
     
     # For manual triggers, check email preferences
@@ -355,7 +353,7 @@ def process_email_notifications(all_warranties, eligible_user_ids, is_manual, ge
             server.login(smtp_username, smtp_password)
 
         emails_sent = 0
-        utc_now = datetime.now(UTC)
+        utc_now = datetime.utcnow()
         timestamp = int(utc_now.timestamp())
         
         for email, user_data in users_warranties.items():
@@ -537,7 +535,7 @@ def send_expiration_notifications(manual_trigger=False, get_db_connection=None, 
 
         if not manual_trigger:
             with conn.cursor() as cur:
-                utc_now = datetime.now(UTC)
+                utc_now = datetime.utcnow()
                 
                 # Check if required columns exist for dynamic query building
                 cur.execute("""
