+++ env.example
@@ -27,7 +27,6 @@ OIDC_CLIENT_ID=your_oidc_client_id
 OIDC_CLIENT_SECRET=your_oidc_client_secret
 OIDC_ISSUER_URL=https://your-oidc-provider.com/auth/realms/your-realm
 OIDC_SCOPE=openid email profile
-OIDC_ADMIN_GROUP=admin # Requires group scope
 
 # Memory and Performance Settings
 WARRACKER_MEMORY_MODE=optimized
