+++ frontend/settings-new.js
@@ -65,7 +65,6 @@ const oidcClientIdInput = document.getElementById('oidcClientId');
 const oidcClientSecretInput = document.getElementById('oidcClientSecret');
 const oidcIssuerUrlInput = document.getElementById('oidcIssuerUrl');
 const oidcScopeInput = document.getElementById('oidcScope');
-const oidcAdminGroupInput = document.getElementById('oidcAdminGroup');
 const saveOidcSettingsBtn = document.getElementById('saveOidcSettingsBtn');
 const oidcRestartMessage = document.getElementById('oidcRestartMessage');
 
@@ -616,20 +615,6 @@ async function loadUserData() {
             }
             if (userNameDisplay) userNameDisplay.textContent = displayName;
             if (userEmailDisplay) userEmailDisplay.textContent = currentUser.email || 'N/A';
-
-            if (currentUser.oidc_managed) {
-                if (firstNameInput) firstNameInput.disabled = true;
-                if (lastNameInput) lastNameInput.disabled = true;
-                if (emailInput) emailInput.disabled = true;
-                if (saveProfileBtn) saveProfileBtn.style.display = 'none';
-                userEditDesc = document.querySelector('#currentUserInfoDisplay > p > strong')
-                if (userEditDesc) {
-                    userEditDesc.setAttribute('data-i18n', 'settings.current_user_oidc')
-                    userEditDesc.textContent = 'OIDC managed profile for:'
-                }
-                securitySection = document.getElementById('securitySection');
-                if (securitySection) securitySection.style.display = 'none';
-            }
             // --- END UPDATE ---
 
             // Admin section visibility will be determined after API call
@@ -3341,7 +3326,6 @@ async function loadSiteSettings() {
     const oidcClientSecretInputElem = document.getElementById('oidcClientSecret');
     const oidcIssuerUrlInputElem = document.getElementById('oidcIssuerUrl');
     const oidcScopeInputElem = document.getElementById('oidcScope');
-    const oidcAdminGroupInputElem = document.getElementById('oidcAdminGroup');
 
     try {
         showLoading();
@@ -3445,13 +3429,6 @@ async function loadSiteSettings() {
         } else {
             console.error('[OIDC Settings] oidcScopeInputElem element NOT FOUND locally.');
         }
-
-        if (oidcAdminGroupInputElem) {
-            console.log('[OIDC Settings] Found oidcAdminGroupInputElem. Setting value to:', settings.oidc_admin_group || '');
-            oidcAdminGroupInputElem.value = settings.oidc_admin_group || '';
-        } else {
-            console.error('[OIDC Settings] oidcAdminGroupInputElem element NOT FOUND locally.');
-        }
         
         console.log('Site and OIDC settings loaded and population attempted using locally queried elements.');
         
@@ -3549,7 +3526,6 @@ async function saveOidcSettings() {
         oidc_client_id: oidcClientIdInput ? oidcClientIdInput.value.trim() : '',
         oidc_issuer_url: oidcIssuerUrlInput ? oidcIssuerUrlInput.value.trim() : '',
         oidc_scope: oidcScopeInput ? oidcScopeInput.value.trim() : 'openid email profile',
-        oidc_admin_group: oidcAdminGroupInput ? oidcAdminGroupInput.value.trim() : ''
     };
 
     // Only include client_secret if a new value is entered
