+++ frontend/status.html
@@ -344,12 +344,11 @@
                             </a>
                         </div>
                         <div class="user-menu-item">
-                            <a href="about.html" style="text-decoration: none; color: inherit;">
-                                <i class="fas fa-info-circle"></i> <span data-i18n="nav.about">About</span>
-                            </a>
+                            <i class="fas fa-info-circle"></i>
+                            <a href="about.html" style="text-decoration: none; color: inherit;" data-i18n="nav.about">About</a>
                         </div>
                         <div class="user-menu-item" id="logoutMenuItem">
-                            <span><i class="fas fa-sign-out-alt"></i> <span data-i18n="auth.logout">Logout</span></span>
+                            <i class="fas fa-sign-out-alt"></i> <span data-i18n="auth.logout">Logout</span>
                         </div>
                     </div>
                 </div>
